move unit tests into tests/unit subfolder
